dat.smith {weightr}	R Documentation

Studies From Smith, Glass, and Miller's (1980) Meta-Analysis of Psychotherapy Outcomes

Description

An arbitrary subset of 74 studies from a meta-analysis assessing the effectiveness of psychotherapy. Contains two moderator variables.

Usage

dat.smith

Format

A data frame containing the following columns:

es

standardized mean difference effect sizes

v

corresponding sampling variances

age

continuous moderator representing clients' average age in years

diagnosis

categorical moderator representing disorder for which clients were treated; 1 = complex phobia, 2 = simple phobia, 3 = other

Details

This dataset consists of an arbitrarily selected subset of 74 studies assessing the effectiveness of psychotherapy. Smith, Glass, and Miller (1980) published a meta-analysis designed to explore the current state of knowledge about psychotherapy effectiveness. Their original meta-analysis contains more than 1,700 effect sizes from 475 studies with multiple moderators and outcome measures. This subset is vastly simplified and intended solely for the purpose of demonstration.

Source

Smith, M. L., Glass, G. V., & Miller, T. I. (1980). Meta-analysis of psychotherapy. American Psychologist, 41, 165-180.

References

Smith, M. L., Glass, G. V., & Miller, T. I. (1980). Meta-analysis of psychotherapy. American Psychologist, 41, 165-180.

Examples

## Not run: 
dat.smith
effect <- dat.smith$es
v <- dat.smith$v
weightfunct(effect, v)

## End(Not run)

[Package weightr version 2.0.1 Index]