funnel {weightr}	R Documentation

Create a Funnel Plot

Description

This function allows you to create a funnel plot using a vector of effect sizes and a vector of their corresponding sampling variances.

Usage

funnel(effect, v, type = "se")

Arguments

effect
a vector of meta-analytic effect sizes

v
a vector of sampling variances

type
v for variance or se for standard error; defaults to standard error

Details

This funnel plot, by default, plots the effect sizes on the y-axis and the measure of study size (either variance or standard error) on the x-axis. If no asymmetry is present, the plot should resemble a horizontal funnel.

Users can choose either standard error (default) or sampling variance as a measure of study size. The choice is mostly arbitrary. In both cases, however, v must be a vector of variances, the same as that required by weightfunct. The conversion to standard error is automatic.

Examples

## Not run: 
# Funnel plot using standard error (default):
funnel(effect, v)
# Funnel plot using sampling variance:
funnel(effect, v, type='v')

## End(Not run)

[Package weightr version 2.0.1 Index]